feat: hoist cap files, fix docs paths, clean package.json
