+++ package.json
@@ -13,6 +13,7 @@
     "test": "pnpm -r --if-present run test",
     "check": "pnpm -r --if-present run check",
     "build": "pnpm -r --if-present run build",
+    "depcheck": "pnpm -r exec depcheck",
     "docs": "typedoc --out docs"
   },
   "devDependencies": {
