+++ packages/capabilities/readme.md
@@ -2,16 +2,6 @@
 
 [Capabilities](https://en.wikipedia.org/wiki/Capability-based_security) for interacting with [web3.storage](https://web3.storage)
 
-## Exports
-
-* '@web3-storage/capabilities/space`
-* '@web3-storage/capabilities/store`
-* '@web3-storage/capabilities/top`
-* '@web3-storage/capabilities/types`
-* '@web3-storage/capabilities/upload`
-* '@web3-storage/capabilities/utils`
-* '@web3-storage/capabilities/voucher`
-
 ## Install
 
 Install the package:
@@ -23,3 +13,15 @@ npm install @web3-storage/capabilities
 ## Usage
 
 [API Reference](https://web3-storage.github.io/w3protocol/modules/_web3_storage_capabilities.html)
+
+```js
+import * as Space from '@web3-storage/capabilities/space'
+import * as Store from '@web3-storage/capabilities/store'
+import * as Top from '@web3-storage/capabilities/top'
+import * as Types from '@web3-storage/capabilities/types'
+import * as Upload from '@web3-storage/capabilities/upload'
+import * as Utils from '@web3-storage/capabilities/utils'
+import * as Voucher from '@web3-storage/capabilities/voucher'
+
+// This package has a "main" entrypoint but we recommend the usage of the specific imports above
+```
