+++ packages/capabilities/src/index.js
@@ -1,5 +1,6 @@
-export * as Space from './capabilities/space.js'
-export * as Top from './capabilities/top.js'
-export * as Store from './capabilities/store.js'
-export * as Upload from './capabilities/upload.js'
-export * as Voucher from './capabilities/voucher.js'
+export * as Space from './space.js'
+export * as Top from './top.js'
+export * as Store from './store.js'
+export * as Upload from './upload.js'
+export * as Voucher from './voucher.js'
+export * as Utils from './utils.js'
