+++ packages/capabilities/src/types.ts
