+++ packages/capabilities/tsconfig.json
@@ -2,9 +2,8 @@
   "extends": "../../tsconfig.json",
   "compilerOptions": {
     "outDir": "dist",
-    "lib": ["ESNext", "DOM"],
     "emitDeclarationOnly": true
   },
-  "include": ["src", "scripts", "test", "package.json"],
+  "include": ["src", "test", "package.json"],
   "exclude": ["**/node_modules/**"]
 }
